Fix all C022 (implicit-real-kind) violations and enforce rule
- Change bare real to real(dp) in 4 files with genuine precision bugs:
  initphonons.F90 (HessCp2k, disAtom/dd), rotatedegen.F90 (smallestFraction),
  parser.F90 (x in is_numeric)
- Upgrade timer.F90 from real to real(dp) for all timer components and locals,
  improving timing precision for long simulations
- Add ! allow(C022) suppression comments in allocatablelist.fypp template
  for scaling factor parameters (intentionally single precision)
- Remove C022 from ignore list (12 violations → 0)

+++ app/phonons/initphonons.F90
@@ -1075,7 +1075,7 @@ contains
     integer :: nDerivs, nBlocks
 
     type(TFileDescr) :: fd
-    real, dimension(:,:), allocatable :: HessCp2k
+    real(dp), dimension(:,:), allocatable :: HessCp2k
     integer ::  n, j1, j2,  p,  q
     type(string) :: filename
     logical :: texist
@@ -1489,7 +1489,7 @@ contains
     integer ::  iAtom, jAtom, ii, jj, kk, PL1, PL2
     !* First guess for nr. of neighbors.
     integer, parameter :: nInitNeighbours = 100
-    real :: disAtom, dd(3)
+    real(dp) :: disAtom, dd(3)
     integer :: nAllAtom
     real(dp) :: mCutoff
     real(dp), allocatable :: coords(:,:), cellVec(:,:), rCellVec(:,:)
